+++ src/main/resources/application.properties
@@ -14,3 +14,5 @@ logging.level.org.hibernate=INFO
 spring.jpa.show-sql=false
 spring.jpa.properties.hibernate.format_sql=false
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=INFO
+
+springdoc.swagger-ui.operationsSorter=method
